Готовый listcard, нужно будет разбить на компоненты
